# src/main/java/ru/itmo/music/music_service/api/dto/CreateTrackResponse.java
package ru.itmo.music.music_service.api.dto;

import com.fasterxml.jackson.annotation.JsonInclude;

import java.time.Instant;

@JsonInclude(JsonInclude.Include.NON_NULL)
public record CreateTrackResponse(
        String id,
        Instant createdAt
) {
}