# src/main/java/ru/itmo/music/music_service/api/error/ErrorResponse.java
package ru.itmo.music.music_service.api.error;

import com.fasterxml.jackson.annotation.JsonInclude;

import java.util.Map;

/**
 * Standard error payload for API responses.
 */
@JsonInclude(JsonInclude.Include.NON_NULL)
public record ErrorResponse(
        String error,
        String message,
        Map<String, Object> details
) {
}